Moved weight clipper to end
