+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod model;
 pub mod dataset;
 pub mod training;
 pub mod convertor;
+mod weight_clipper;
